backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0,<5.0.0
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator>=2.0.0
httpx==0.26.0
alembic==1.13.1
scikit-learn>=1.3.0
boto3>=1.34.0
openai>=1.0.0
pypdf>=4.0.0
ruff>=0.1.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.26.0
pytest-cov>=4.1.0
aiosqlite>=0.19.0
